+++ dcim/api.py
@@ -9,16 +9,23 @@ many times as it will not make use of TLS connection pooling and keep-alive.
 from dcim.client import DCIMClient
 
 
-def locate(device):
-    with DCIMClient() as client:
-        return client.locate(device)
+def locate(*args, **kwargs):
+    with DCIMClient(caching=True) as client:
+        return client.locate(*args, **kwargs)
 
 locate.__doc__ = DCIMClient.locate.__doc__
 
 
-def showrack(location, display=False, width=72):
-    with DCIMClient() as client:
-        return client.showrack(location, display=display, width=width)
+def model(*args, **kwargs):
+    with DCIMClient(caching=True) as client:
+        return client.model(*args, **kwargs)
+
+model.__doc__ = DCIMClient.model.__doc__
+
+
+def showrack(*args, **kwargs):
+    with DCIMClient(caching=True) as client:
+        return client.showrack(*args, **kwargs)
 
 showrack.__doc__ = DCIMClient.showrack.__doc__