+++ CHANGES.rst
@@ -1,6 +1,9 @@
 Unreleased
 ==========
 
+- [NEW] Add model command to show make, model, and serial of a device
+- [NEW] Add option to showrack to display make, model, and serial of each
+        device.
 - [IMPROVED] Add GET response caching for performance/testing
 
 0.0.1 (2017-09-13)